refactor Join namespace
